requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
Django==3.2.4
django-environ==0.4.5
django-storages==1.11.1
djangorestframework==3.12.4
environ==1.0
google-api-core==1.29.0
google-auth==1.30.1
google-cloud-core==1.6.0
google-cloud-secret-manager==2.5.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.38.0
gunicorn==20.1.0
idna==2.10
libcst==0.3.19
mypy-extensions==0.4.3
mysqlclient==2.0.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.5
